script: another edit

+++ yumi.sh
@@ -6,6 +6,15 @@ export ANDROID_MAJOR_VERSION=q
 export ANDROID_PLATFORM_VERSION=10
 export CCACHE_DIR=/target/.ccache
 
+read -p "Clean source (y/n) > " yn
+if [ "$yn" = "Y" -o "$yn" = "y" ]; then
+     echo "Clean Build"
+     CLEAN="1"
+else
+     echo "Dirty Build"
+     CLEAN="0"
+fi
+
 for device in $*; do
   case $device in
     A320X) DVNAME=exynos7870-a3y17lte_defconfig
@@ -26,17 +35,21 @@ for device in $*; do
 	   ;;
   esac
   echo Building kernel for $device...
-  make clean && make mrproper ;
-  rm -rf ./Yumi/Out ;
+  if [ $CLEAN = 1 ] ; then
+	echo Cleaning build dir
+  	make clean && make mrproper ;
+  	rm -rf ./Yumi/Out ;
+  fi
   make O=./Yumi/Out $DVNAME ;
   make O=./Yumi/Out -j$(nproc --all) ;
-  sudo cp -r ./Yumi/Q/ramdisk ./Yumi/AIK/ramdisk ;
-  cp -r ./Yumi/Q/split_img ./Yumi/AIK/split_img ;
-  mv ./Yumi/Out/arch/arm64/boot/Image ./Yumi/AIK/split_img/boot.img-zImage ;
-  mv ./Yumi/Out/arch/arm64/boot/dtb.img ./Yumi/AIK/split_img/boot.img-dt ;
-  sudo ./Yumi/AIK/repackimg.sh ;
+  ./Yumi/AIK-Linux/cleanup.sh
+  sudo cp -r ./Yumi/Q/ramdisk ./Yumi/AIK-Linux/ramdisk ;
+  cp -r ./Yumi/Q/split_img ./Yumi/AIK-Linux/split_img ;
+  mv ./Yumi/Out/arch/arm64/boot/Image ./Yumi/AIK-Linux/split_img/boot.img-zImage ;
+  mv ./Yumi/Out/arch/arm64/boot/dtb.img ./Yumi/AIK-Linux/split_img/boot.img-dt ;
+  sudo ./Yumi/AIK-Linux/repackimg.sh ;
   mkdir ./Yumi/Product
-  mv ./Yumi/AIK/image-new.img ./Yumi/Product/YumiKernel-V1.0-$device-$(date +'%Y%m%d').img ;
-  ./Yumi/AIK/cleanup.sh ;
+  mv ./Yumi/AIK-Linux/image-new.img ./Yumi/Product/YumiKernel-V1.0-$device-$(date +'%Y%m%d').img ;
+  ./Yumi/AIK-Linux/cleanup.sh ;
   echo Kernel build complete for $device
 done
